move hash to its own file, similarly to navigate tests
